agent-compose/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agent-compose"
version = "0.1.0"
description = "Docker Compose for multi-agent AI systems"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Daniel López Orta", email = "daniellopezorta39@gmail.com" }]
keywords = [
    "agent", "multi-agent", "orchestration", "langgraph", "crewai",
    "yaml", "compose", "pipeline", "llm",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.30"]
langgraph = ["langgraph>=0.2", "langchain-openai>=0.2"]
crewai = ["crewai>=0.80"]
all = [
    "openai>=1.0",
    "anthropic>=0.30",
    "langgraph>=0.2",
    "langchain-openai>=0.2",
    "crewai>=0.80",
    "tiktoken>=0.7.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
agent-compose = "agent_compose.cli:main"

[project.urls]
Homepage = "https://github.com/daniellopez882/agent-compose"
Repository = "https://github.com/daniellopez882/agent-compose"
Issues = "https://github.com/daniellopez882/agent-compose/issues"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
